requirements.txt
certifi==2018.11.29
chardet==3.0.4
Cython==0.29.3
docutils==0.14
idna==2.8
Kivy==1.10.1
Kivy-Garden==0.1.4
Pygments==2.3.1
pyserial==3.4
requests==2.21.0
urllib3==1.24.1
webcolors==1.8.1
